+++ microserv-billing/src/main/java/org/example/microservbilling/services/FeeService.java
@@ -3,6 +3,7 @@ import org.example.microservbilling.dto.BillingDto;
 import org.example.microservbilling.dto.FeeDto;
 import org.example.microservbilling.entity.Billing;
 import org.example.microservbilling.entity.Fee;
+import org.example.microservbilling.error.exception.NotExistsException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.example.microservbilling.repository.FeeRepository;
@@ -34,13 +35,17 @@ public class FeeService {
     //devuelve una tarifa por id
     public FeeDto getFee(Long id){
         Optional<Fee>feeOptional=feeRepository.findById(id);
-        Fee fee = feeOptional.get();
+        if(feeOptional.isPresent()){
         return FeeDto.builder()
-                .id(fee.getId())
-                .monto(fee.getMonto())
-                .fechaInicio(fee.getFechaInicio())
-                .tipo(fee.getTipo()).build();
+                .id(feeOptional.get().getId())
+                .monto(feeOptional.get().getMonto())
+                .fechaInicio(feeOptional.get().getFechaInicio())
+                .tipo(feeOptional.get().getTipo()).build();
+        }else {
+        throw new NotExistsException("El id: " + id + " No existe");
+        }
     }
+
 
     //crea una tarifa
     public FeeDto createFee(Fee newFee){
@@ -52,7 +57,16 @@ public class FeeService {
                 .tipo(fee.getTipo()).build();
     }
     //elimina una tarifa
-    public void delteFee(Long id){
-        feeRepository.deleteById(id);
+    public FeeDto delteFee(Long id){
+        Optional<Fee> feeOptional = feeRepository.findById(id);
+        if(feeOptional.isPresent()) {
+            feeRepository.deleteById(id);
+            return FeeDto.builder().id(feeOptional.get().getId())
+                                    .fechaInicio(feeOptional.get().getFechaInicio())
+                                    .monto(feeOptional.get().getMonto())
+                                    .tipo(feeOptional.get().getTipo()).build();
+        }else {
+            throw new NotExistsException("El id que quieres eliminar no existe. ID: " +id);
+        }
     }
 }
